updateMoments -> displayMoment
This commit clears up some confusion that I had around the scope
and name of things. `currentMoment` is now the keeper of state;
`randomMoment`, `nextMoment`, `prevMoment` update that state and
then call out to `displayMoment`, passing it the value of
`currentMoment`.

This means that `displayMoment` could *almost* live entirely
outside of this file; if you also passed it all the data, it could
be entirely separate. You can even start to think of the
`displayMoment` function as representing your "view," and the state
is your "model." The model is where you change state; the view
is just a visual representation of that state, but doesn't
change it ... at least not directly.